Refactoring QGCFlightDisplay and Renaming QGCMavManager.* Sorting source and headers while at it.
